requirements.txt
pandas>=1.1.5
requests>=2.25.1
python-dotenv==1.0.0